refactor(cfd-validation): Decompose reporting/mod.rs into leaf modules
Part of ATLAS-ARCH-007.
