+++ crates/cfd-validation/src/reporting/mod.rs
@@ -6,632 +6,25 @@
 //! - Convergence studies and error analysis
 //! - Code quality metrics
 
+mod automation;
+mod builder;
+mod data;
+#[cfg(test)]
+mod tests;
+
 pub mod fidelity_limits;
 pub mod html;
 pub mod json;
 pub mod markdown;
 pub mod summary;
 
+pub use automation::AutomatedReporter;
+pub use builder::ReportBuilder;
+pub use data::{
+    CodeQualityReport, PerformanceReport, Reporter, TestCategory, TestResult, TestStatus,
+    ValidationReport,
+};
 pub use html::HtmlReporter;
 pub use json::JsonReporter;
 pub use markdown::MarkdownReporter;
 pub use summary::{PerformanceMetrics, ValidationSummary};
-
-use cfd_core::error::Result;
-use std::collections::HashMap;
-use std::time::SystemTime;
-
-/// Main validation report structure
-#[derive(Debug, Clone)]
-pub struct ValidationReport {
-    /// Report generation timestamp
-    pub timestamp: SystemTime,
-    /// Report title
-    pub title: String,
-    /// Executive summary
-    pub summary: ValidationSummary,
-    /// Test results by category
-    pub test_results: HashMap<String, TestCategory>,
-    /// Performance benchmarks
-    pub performance: Vec<PerformanceReport>,
-    /// Code quality metrics
-    pub code_quality: CodeQualityReport,
-    /// Recommendations for improvement
-    pub recommendations: Vec<String>,
-}
-
-/// Test category results
-#[derive(Debug, Clone)]
-pub struct TestCategory {
-    /// Name of the test category
-    pub name: String,
-    /// Number of passed tests
-    pub passed: usize,
-    /// Number of failed tests
-    pub failed: usize,
-    /// Number of skipped tests
-    pub skipped: usize,
-    /// Total number of tests
-    pub total: usize,
-    /// Test coverage percentage
-    pub coverage_percentage: f64,
-    /// Detailed test results
-    pub details: Vec<TestResult>,
-}
-
-/// Individual test result
-#[derive(Debug, Clone)]
-pub struct TestResult {
-    /// Name of the test
-    pub name: String,
-    /// Execution status
-    pub status: TestStatus,
-    /// Execution duration in milliseconds
-    pub duration_ms: f64,
-    /// Error message if failed
-    pub error_message: Option<String>,
-    /// Coverage data if available
-    pub coverage_data: Option<String>,
-}
-
-/// Test execution status
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub enum TestStatus {
-    /// Test passed
-    Passed,
-    /// Test failed
-    Failed,
-    /// Test was skipped
-    Skipped,
-    /// Test timed out
-    Timeout,
-}
-
-/// Performance benchmark report
-#[derive(Debug, Clone)]
-pub struct PerformanceReport {
-    /// Name of the benchmark
-    pub benchmark_name: String,
-    /// Performance metrics
-    pub metrics: PerformanceMetrics,
-    /// Percentage change if regression detected
-    pub regression_detected: Option<f64>, // percentage change
-    /// Baseline metrics for comparison
-    pub baseline_comparison: Option<PerformanceMetrics>,
-}
-
-/// Code quality report
-#[derive(Debug, Clone)]
-pub struct CodeQualityReport {
-    /// Total lines of code
-    pub lines_of_code: usize,
-    /// Test coverage percentage
-    pub test_coverage: f64,
-    /// Documentation coverage percentage
-    pub documentation_coverage: f64,
-    /// Number of clippy warnings
-    pub clippy_warnings: usize,
-    /// Number of compiler errors
-    pub compiler_errors: usize,
-    /// Cyclomatic complexity score
-    pub cyclomatic_complexity: f64,
-    /// Maintainability index
-    pub maintainability_index: f64,
-}
-
-/// Validation report builder
-pub struct ReportBuilder {
-    title: String,
-    summary: ValidationSummary,
-    test_results: HashMap<String, TestCategory>,
-    performance: Vec<PerformanceReport>,
-    code_quality: CodeQualityReport,
-    recommendations: Vec<String>,
-}
-
-impl ReportBuilder {
-    /// Create a new report builder with the specified title
-    pub fn new(title: String) -> Self {
-        Self {
-            title,
-            summary: ValidationSummary::default(),
-            test_results: HashMap::new(),
-            performance: Vec::new(),
-            code_quality: CodeQualityReport::default(),
-            recommendations: Vec::new(),
-        }
-    }
-
-    /// Set the validation summary
-    pub fn with_summary(mut self, summary: ValidationSummary) -> Self {
-        self.summary = summary;
-        self
-    }
-
-    /// Add a test category result
-    pub fn add_test_category(mut self, category: TestCategory) -> Self {
-        self.test_results.insert(category.name.clone(), category);
-        self
-    }
-
-    /// Add a performance benchmark report
-    pub fn add_performance_report(mut self, report: PerformanceReport) -> Self {
-        self.performance.push(report);
-        self
-    }
-
-    /// Set code quality metrics
-    pub fn with_code_quality(mut self, quality: CodeQualityReport) -> Self {
-        self.code_quality = quality;
-        self
-    }
-
-    /// Add a recommendation for improvement
-    pub fn add_recommendation(mut self, recommendation: String) -> Self {
-        self.recommendations.push(recommendation);
-        self
-    }
-
-    /// Build the final validation report
-    pub fn build(self) -> ValidationReport {
-        ValidationReport {
-            timestamp: SystemTime::now(),
-            title: self.title,
-            summary: self.summary,
-            test_results: self.test_results,
-            performance: self.performance,
-            code_quality: self.code_quality,
-            recommendations: self.recommendations,
-        }
-    }
-}
-
-impl Default for CodeQualityReport {
-    fn default() -> Self {
-        Self {
-            lines_of_code: 0,
-            test_coverage: 0.0,
-            documentation_coverage: 0.0,
-            clippy_warnings: 0,
-            compiler_errors: 0,
-            cyclomatic_complexity: 0.0,
-            maintainability_index: 0.0,
-        }
-    }
-}
-
-impl ValidationReport {
-    /// Generate report in specified format
-    pub fn generate<T: Reporter>(&self, reporter: &T) -> Result<String> {
-        reporter.generate_report(self)
-    }
-
-    /// Calculate overall health score (0.0 to 1.0)
-    pub fn health_score(&self) -> f64 {
-        let test_score = if self.summary.total_tests > 0 {
-            self.summary.passed_tests as f64 / self.summary.total_tests as f64
-        } else {
-            0.0
-        };
-
-        let coverage_score = self.code_quality.test_coverage / 100.0;
-        let quality_score = 1.0 - (self.code_quality.clippy_warnings as f64 * 0.01).min(1.0);
-
-        // Weighted average
-        0.4 * test_score + 0.4 * coverage_score + 0.2 * quality_score
-    }
-
-    /// Get critical issues requiring attention
-    pub fn critical_issues(&self) -> Vec<String> {
-        let mut issues = Vec::new();
-
-        if self.summary.failed_tests > 0 {
-            issues.push(format!(
-                "{} test failures detected",
-                self.summary.failed_tests
-            ));
-        }
-
-        if self.code_quality.test_coverage < 80.0 {
-            issues.push(format!(
-                "Test coverage below 80%: {:.1}%",
-                self.code_quality.test_coverage
-            ));
-        }
-
-        if self.code_quality.clippy_warnings > 10 {
-            issues.push(format!(
-                "High number of clippy warnings: {}",
-                self.code_quality.clippy_warnings
-            ));
-        }
-
-        let regressions: Vec<_> = self
-            .performance
-            .iter()
-            .filter(|p| p.regression_detected.is_some())
-            .collect();
-
-        if !regressions.is_empty() {
-            issues.push(format!(
-                "{} performance regressions detected",
-                regressions.len()
-            ));
-        }
-
-        issues
-    }
-}
-
-/// Reporter trait for different output formats
-pub trait Reporter {
-    /// Generate report string from validation report data
-    fn generate_report(&self, report: &ValidationReport) -> Result<String>;
-}
-
-/// Automated report generation utilities
-pub struct AutomatedReporter;
-
-impl AutomatedReporter {
-    /// Generate comprehensive validation report from test results
-    pub fn generate_from_tests(
-        test_output: &str,
-        _coverage_data: Option<&str>,
-    ) -> Result<ValidationReport> {
-        let mut builder = ReportBuilder::new("CFD Validation Report".to_string());
-
-        // Parse real cargo test output into structured ValidationSummary
-        let summary = Self::parse_test_summary(test_output)?;
-        builder = builder.with_summary(summary);
-
-        // Parse test output to extract categories and details
-        let test_categories = Self::extract_test_categories(test_output)?;
-
-        for category in test_categories {
-            builder = builder.add_test_category(category);
-        }
-
-        // Add code quality metrics from real tooling outputs
-        let code_quality = Self::derive_code_quality_metrics(test_output)?;
-        builder = builder.with_code_quality(code_quality);
-
-        // Add recommendations
-        builder = builder
-            .add_recommendation("Increase test coverage to >90%".to_string())
-            .add_recommendation("Add performance regression tests".to_string())
-            .add_recommendation("Implement automated documentation generation".to_string());
-
-        Ok(builder.build())
-    }
-
-    /// Parse test summary from cargo test output
-    fn parse_test_summary(output: &str) -> Result<ValidationSummary> {
-        let lines: Vec<&str> = output.lines().collect();
-
-        let mut passed_tests = 0;
-        let mut failed_tests = 0;
-        let mut ignored_tests = 0;
-        let mut measured_tests = 0;
-        let mut filtered_tests = 0;
-        let mut total_duration_secs = 0.0;
-
-        // Parse comprehensive test results
-        for line in lines {
-            // Parse test result summary line
-            if line.contains("test result:") {
-                // Handle both success and failure cases
-                if line.contains("ok") || line.contains("FAILED") {
-                    // Extract numbers from patterns like:
-                    // "test result: ok. 10 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out"
-                    // "test result: FAILED. 9 passed; 1 failed; 0 ignored; 0 measured; 0 filtered out"
-
-                    let clean_line = line
-                        .replace("test result:", "")
-                        .replace("ok.", "")
-                        .replace("FAILED.", "");
-
-                    for part in clean_line.split(';') {
-                        let part = part.trim();
-                        if part.contains("passed") {
-                            if let Some(num_str) = part.split_whitespace().next() {
-                                if let Ok(num) = num_str.parse::<usize>() {
-                                    passed_tests = num;
-                                }
-                            }
-                        } else if part.contains("failed") {
-                            if let Some(num_str) = part.split_whitespace().next() {
-                                if let Ok(num) = num_str.parse::<usize>() {
-                                    failed_tests = num;
-                                }
-                            }
-                        } else if part.contains("ignored") {
-                            if let Some(num_str) = part.split_whitespace().next() {
-                                if let Ok(num) = num_str.parse::<usize>() {
-                                    ignored_tests = num;
-                                }
-                            }
-                        } else if part.contains("measured") {
-                            if let Some(num_str) = part.split_whitespace().next() {
-                                if let Ok(num) = num_str.parse::<usize>() {
-                                    measured_tests = num;
-                                }
-                            }
-                        } else if part.contains("filtered") {
-                            if let Some(num_str) = part.split_whitespace().next() {
-                                if let Ok(num) = num_str.parse::<usize>() {
-                                    filtered_tests = num;
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-
-            // Parse duration from lines like "test ... ... ok 1.23s" or "test ... ... FAILED 0.45s"
-            if (line.contains(" ok ") || line.contains(" FAILED ")) && line.contains('s') {
-                let parts: Vec<&str> = line.split_whitespace().collect();
-                if let Some(last_part) = parts.last() {
-                    if let Ok(duration) = last_part.trim_end_matches('s').parse::<f64>() {
-                        total_duration_secs += duration;
-                    }
-                }
-            }
-        }
-
-        let total_tests =
-            passed_tests + failed_tests + ignored_tests + measured_tests + filtered_tests;
-
-        Ok(ValidationSummary {
-            total_tests,
-            passed_tests,
-            failed_tests,
-            skipped_tests: ignored_tests + filtered_tests,
-            total_duration: std::time::Duration::from_secs_f64(total_duration_secs),
-            coverage_percentage: if total_tests > 0 {
-                (passed_tests as f64 / total_tests as f64) * 100.0
-            } else {
-                0.0
-            },
-        })
-    }
-
-    /// Extract test categories from cargo test output
-    fn extract_test_categories(test_output: &str) -> Result<Vec<TestCategory>> {
-        let mut categories = std::collections::HashMap::new();
-
-        // Parse individual test results to categorize them
-        for line in test_output.lines() {
-            if line.starts_with("test ") && (line.contains(" ok ") || line.contains(" FAILED ")) {
-                // Extract test name and categorize based on naming patterns
-                let parts: Vec<&str> = line.split_whitespace().collect();
-                if parts.len() >= 3 {
-                    let test_name = parts[1];
-
-                    // Determine category based on test name patterns
-                    let category = if test_name.contains("mms")
-                        || test_name.contains("manufactured")
-                    {
-                        "MMS Validation"
-                    } else if test_name.contains("turbulent")
-                        || test_name.contains("k_omega")
-                        || test_name.contains("reynolds")
-                    {
-                        "Turbulent Flow"
-                    } else if test_name.contains("ghia") || test_name.contains("cavity") {
-                        "Benchmark Validation"
-                    } else if test_name.contains("poiseuille") || test_name.contains("channel") {
-                        "Fundamental Flows"
-                    } else if test_name.contains("richardson") || test_name.contains("convergence")
-                    {
-                        "Convergence Analysis"
-                    } else {
-                        "General Tests"
-                    };
-
-                    let entry =
-                        categories
-                            .entry(category.to_string())
-                            .or_insert((0, 0, 0, Vec::new()));
-
-                    let (passed, failed, _skipped, details) = entry;
-                    *passed += usize::from(line.contains(" ok "));
-                    *failed += usize::from(line.contains(" FAILED "));
-
-                    // Add test detail
-                    let status = if line.contains(" ok ") {
-                        TestStatus::Passed
-                    } else {
-                        TestStatus::Failed
-                    };
-                    let duration_ms = if let Some(last_part) = parts.last() {
-                        last_part
-                            .trim_end_matches('s')
-                            .parse::<f64>()
-                            .unwrap_or(0.0)
-                            * 1000.0
-                    } else {
-                        0.0
-                    };
-
-                    details.push(TestResult {
-                        name: test_name.to_string(),
-                        status,
-                        duration_ms,
-                        error_message: None,
-                        coverage_data: None,
-                    });
-                }
-            }
-        }
-
-        // Convert to TestCategory structs
-        let mut result = Vec::new();
-        for (name, (passed, failed, skipped, details)) in categories {
-            let total = passed + failed;
-            let coverage_percentage = if total > 0 {
-                (passed as f64 / total as f64) * 100.0
-            } else {
-                0.0
-            };
-
-            result.push(TestCategory {
-                name,
-                passed,
-                failed,
-                skipped,
-                total,
-                coverage_percentage,
-                details,
-            });
-        }
-
-        Ok(result)
-    }
-
-    /// Derive code quality metrics from test output and available tooling
-    fn derive_code_quality_metrics(test_output: &str) -> Result<CodeQualityReport> {
-        // Extract timing information for performance metrics
-        let _total_duration = Self::extract_total_duration(test_output);
-
-        // Parse compiler warnings from output
-        let _compiler_warnings = test_output
-            .lines()
-            .filter(|line| line.contains("warning:") || line.contains("warning:"))
-            .count();
-
-        // Parse compiler errors from output
-        let compiler_errors = test_output
-            .lines()
-            .filter(|line| line.contains("error:") || line.contains("error:"))
-            .count();
-
-        // For now, use reasonable defaults for metrics that require external tools
-        // In a full implementation, these would be derived from actual tool outputs
-        Ok(CodeQualityReport {
-            lines_of_code: 15420, // Would be derived from `wc -l` or similar
-            test_coverage: Self::calculate_test_coverage(test_output),
-            documentation_coverage: 73.2, // Would be derived from documentation analysis
-            clippy_warnings: 3,           // Would be derived from `cargo clippy` output
-            compiler_errors,
-            cyclomatic_complexity: 2.1, // Would be derived from complexity analysis tools
-            maintainability_index: 78.5, // Would be derived from maintainability analysis
-        })
-    }
-
-    /// Extract total test duration from output
-    fn extract_total_duration(test_output: &str) -> f64 {
-        let mut total_duration = 0.0;
-
-        for line in test_output.lines() {
-            if line.contains("test result:") {
-                // Look for duration in the summary line like "finished in 1.23s"
-                if let Some(duration_str) = line.split("finished in ").nth(1) {
-                    if let Some(duration) = duration_str.split_whitespace().next() {
-                        if let Ok(duration) = duration.trim_end_matches('s').parse::<f64>() {
-                            total_duration = duration;
-                            break;
-                        }
-                    }
-                }
-            }
-        }
-
-        total_duration
-    }
-
-    /// Calculate test coverage percentage from parsed results
-    fn calculate_test_coverage(test_output: &str) -> f64 {
-        let mut total_tests = 0;
-        let mut passed_tests = 0;
-
-        for line in test_output.lines() {
-            if line.contains("test result:") && (line.contains("ok") || line.contains("FAILED")) {
-                // Parse the test summary
-                let clean_line = line
-                    .replace("test result:", "")
-                    .replace("ok.", "")
-                    .replace("FAILED.", "");
-
-                for part in clean_line.split(';') {
-                    let part = part.trim();
-                    if part.contains("passed") {
-                        if let Some(num_str) = part.split_whitespace().next() {
-                            if let Ok(num) = num_str.parse::<usize>() {
-                                passed_tests = num;
-                            }
-                        }
-                    } else if part.contains("failed") {
-                        if let Some(num_str) = part.split_whitespace().next() {
-                            if let Ok(num) = num_str.parse::<usize>() {
-                                total_tests += num;
-                            }
-                        }
-                    }
-                }
-
-                total_tests += passed_tests;
-            }
-        }
-
-        if total_tests > 0 {
-            (passed_tests as f64 / total_tests as f64) * 100.0
-        } else {
-            0.0
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_report_builder() {
-        let report = ReportBuilder::new("Test Report".to_string())
-            .with_summary(ValidationSummary {
-                total_tests: 10,
-                passed_tests: 9,
-                failed_tests: 1,
-                skipped_tests: 0,
-                total_duration: std::time::Duration::from_secs(5),
-                coverage_percentage: 85.0,
-            })
-            .add_recommendation("Fix failing test".to_string())
-            .build();
-
-        assert_eq!(report.title, "Test Report");
-        assert_eq!(report.summary.passed_tests, 9);
-        assert_eq!(report.summary.failed_tests, 1);
-        assert_eq!(report.recommendations.len(), 1);
-    }
-
-    #[test]
-    fn test_health_score() {
-        let report = ValidationReport {
-            timestamp: SystemTime::now(),
-            title: "Test".to_string(),
-            summary: ValidationSummary {
-                total_tests: 10,
-                passed_tests: 8,
-                failed_tests: 2,
-                skipped_tests: 0,
-                total_duration: std::time::Duration::from_secs(1),
-                coverage_percentage: 80.0,
-            },
-            test_results: HashMap::new(),
-            performance: Vec::new(),
-            code_quality: CodeQualityReport {
-                test_coverage: 80.0,
-                clippy_warnings: 5,
-                ..Default::default()
-            },
-            recommendations: Vec::new(),
-        };
-
-        let score = report.health_score();
-        assert!((0.0..=1.0).contains(&score));
-
-        let issues = report.critical_issues();
-        assert!(!issues.is_empty()); // Should detect failures and low coverage
-    }
-}
